refactor: move murcko_scaffold.py to core/chemistry/scaffolds.py
